add calulcateAverage method and calculateAverageTest

+++ src/main/java/pl/paulina/Subject/Subject.java
@@ -14,7 +14,7 @@ public class Subject {
     private  Grade grade;
 
 
-    public ArrayList<Double> grades = new ArrayList<>();
+    public ArrayList<Integer> grades = new ArrayList<>();
 
     public Subject() {}
